requirements.txt
colorlog~=6.8.2
PyYAML~=6.0.1
internal~=0.1.74
Flask~=3.0.2
pandas~=2.2.0
scikit-learn~=1.4.1.post1